feat: 增加域名以及证书信息的查询能力
